Makefile: spl: add rules to generate FIT images for SPL DTB
Some platforms require more complex SPL images that a firmware
needs to be loaded as part of SPL. TI's K3 AM654 SoC is one
such platform. A firmware needs to be loaded to its micro controller
which acts as a central resource manager.

Introduce a CONFIG_SPL_DTB_FIT_SOURCE and CONFIG_SPL_DTB_FIT_GENERATOR
symbol, which can either hold an .its source file describing the image
layout, or, in the second case, a generator tool (script) to create such
a source file. This is inspired by the configs CONFIG_SPL_FIT_SOURCE
and CONFIG_SPL_FIT_GENERATOR which are used to create images for u-boot.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>