firmware: ti_sci: Add support for processor control services
TI-SCI message protocol provides support for controlling of various
physical cores available in SoC. In order to control which host is
capable of controlling a physical processor core, there is a processor
access control list that needs to be populated as part of the board
configuration data.

Introduce support for the set of TI-SCI message protocol apis that
provide us with this capability of controlling physical cores.

ToDo:
- Create separate TISCI commands for control, config, status flags.
- Update Documentation.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>